[37] Refactor Blocks. Gallery block.
