+++ src/blocks/gallery/block.js
@@ -1,20 +1,16 @@
 /**
  * External dependencies
  */
-const { filter, every, get } = lodash;
+
 import './style.scss';
-import Edit, { defaultColumnsNumber } from './edit';
+import {GallerySaveComponent} from "./components/GallerySaveComponent";
+import GalleryEditComponent from "./components/GalleryEditComponent";
 
 /**
  * WordPress dependencies
  */
-
-const { __ } = wp.i18n;
-const el = wp.element.createElement;
-const { registerBlockType } = wp.blocks;
-const RichText = wp.blockEditor.RichText;
-const editorMediaUpload = wp.editor.editorMediaUpload;
-const name = 'core/gallery';
+import { __ } from '@wordpress/i18n';
+import { registerBlockType } from '@wordpress/blocks';
 
 registerBlockType( 'alps-gutenberg-blocks/gallery', {
   title: __('ALPS Gallery', 'alps-gutenberg-blocks'),
@@ -63,38 +59,6 @@ registerBlockType( 'alps-gutenberg-blocks/gallery', {
     },
   },
 
-  edit: function( props ) {
-  	return <Edit {...props} />
-  },
-
-  save: function( { attributes } ) {
-    const { images } = attributes;
-    return (
-      <div className="c-gallery-block__image">
-        <div className="js-this c-gallery-block c-block u-background-color--gray--light u-border--left u-theme--border-color--darker--left can-be--dark-dark">
-          <div className="c-gallery-block__header">
-            <div className="c-gallery-block__title u-padding u-spacing--half">
-              <h2 className="u-font--primary--s u-theme--color--darker">
-                <span className="u-theme--color--base"><em>{ __('Gallery', 'alps-gutenberg-blocks') }</em></span> <span className="o-title">{ attributes.title }</span>
-              </h2>
-              <button className="c-gallery-block__toggle js-toggle o-button o-button--outline o-button--toggle o-button--small" data-toggled="this" data-prefix="this"><span class="u-icon u-icon--xs u-path-fill--white"><svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 10 10"><title>o-icon__plus</title><path d="M10,4H6V0H4V4H0V6H4v4H6V6h4Z" fill="#9b9b9b"/></svg></span></button>
-            </div>
-            <div className="c-gallery-block__thumb u-background--contain" style={ { backgroundImage: `url('${ images.map( ( image, index ) => image.url )[0] }');` } }>
-            </div>
-          </div>
-
-          <div className="c-gallery-block__body">
-            { images.map( ( image ) =>
-              <div key={ image.id || image.url } className="c-gallery-block__image">
-                <img src={ image.url } alt={ image.alt } data-id={ image.id } />
-                <div className="c-gallery-block__caption u-font--secondary--s u-color--gray u-padding u-padding--double--bottom">
-                  { image.caption }
-                </div>
-              </div>
-            ) }
-          </div>
-        </div>
-      </div>
-    );
-  },
+  edit: GalleryEditComponent,
+  save: GallerySaveComponent
 } );
